Convert projects to use a nullable context.

+++ src/Stott.Security.Optimizely/Features/Permissions/Service/CspPermissionService.cs
@@ -9,7 +9,7 @@ using Stott.Security.Optimizely.Entities;
 using Stott.Security.Optimizely.Features.Caching;
 using Stott.Security.Optimizely.Features.Permissions.Repository;
 
-public class CspPermissionService : ICspPermissionService
+internal sealed class CspPermissionService : ICspPermissionService
 {
     private readonly ICspPermissionRepository _repository;
 
@@ -23,15 +23,25 @@ public class CspPermissionService : ICspPermissionService
         _cacheWrapper = cacheWrapper ?? throw new ArgumentNullException(nameof(cacheWrapper));
     }
 
-    public async Task AppendDirectiveAsync(string source, string directive, string modifiedBy)
+    public async Task AppendDirectiveAsync(string? source, string? directive, string? modifiedBy)
     {
+        if (string.IsNullOrWhiteSpace(source) || string.IsNullOrWhiteSpace(directive) || string.IsNullOrWhiteSpace(modifiedBy))
+        {
+            return;
+        }
+
         await _repository.AppendDirectiveAsync(source, directive, modifiedBy);
 
         _cacheWrapper.Remove(CspConstants.CacheKeys.CompiledCsp);
     }
 
-    public async Task DeleteAsync(Guid id, string deletedBy)
+    public async Task DeleteAsync(Guid id, string? deletedBy)
     {
+        if (id.Equals(Guid.Empty) || string.IsNullOrWhiteSpace(deletedBy))
+        {
+            return;
+        }
+
         await _repository.DeleteAsync(id, deletedBy);
 
         _cacheWrapper.Remove(CspConstants.CacheKeys.CompiledCsp);
@@ -47,8 +57,13 @@ public class CspPermissionService : ICspPermissionService
         return _repository.GetCmsRequirements();
     }
 
-    public async Task SaveAsync(Guid id, string source, List<string> directives, string modifiedBy)
+    public async Task SaveAsync(Guid id, string? source, List<string>? directives, string? modifiedBy)
     {
+        if (string.IsNullOrWhiteSpace(source) || string.IsNullOrWhiteSpace(modifiedBy) || directives is not { Count: > 0 })
+        {
+            return;
+        }
+
         await _repository.SaveAsync(id, source, directives, modifiedBy);
 
         _cacheWrapper.Remove(CspConstants.CacheKeys.CompiledCsp);
